Mage enemy, several bugs fixed

+++ Assets/_Scripts/DataSO/RuneDataSO.cs
@@ -33,7 +33,7 @@ public class RuneDataSO : ScriptableObject
 
     [field: SerializeField]
     [field: Range(0, 30)]
-    public float PiercingBonus { get; set; } = 0;
+    public int PiercingBonus { get; set; } = 0;
 
     [field: SerializeField]
     [field: Range(0, 30)]
